Make the interval skew configurable
Previously the verify method would only check the current and the previous interval, which causes problems when the user has positive clock skew.

Example:
- Server time is 10:29:00.
- User time is 10:31:00.
- Server will check intervals 09:59:30-10:00:00,10:00:00-10:30:00, and reject the code with only a couple of seconds skew.

This commit makes the default to check both the previous, the current, and the next interval, and also lets the user change this behavior if desired.